+++ test/src/models/membership_test.dart
@@ -11,6 +11,7 @@ void main() {
         userId: '5e5ea5c16897e',
         userName: 'John Doe',
         userEmail: 'john@appwrite.io',
+        userPhone: '+1 555 555 5555',
         teamId: '5e5ea5c16897e',
         teamName: 'VIP',
         invited: '2020-10-15T06:38:00.000+00:00',
@@ -23,19 +24,20 @@ void main() {
       final map = model.toMap();
       final result = Membership.fromMap(map);
 
-      expect(result.$id, '5e5ea5c16897e');
-      expect(result.$createdAt, '2020-10-15T06:38:00.000+00:00');
-      expect(result.$updatedAt, '2020-10-15T06:38:00.000+00:00');
-      expect(result.userId, '5e5ea5c16897e');
-      expect(result.userName, 'John Doe');
-      expect(result.userEmail, 'john@appwrite.io');
-      expect(result.teamId, '5e5ea5c16897e');
-      expect(result.teamName, 'VIP');
-      expect(result.invited, '2020-10-15T06:38:00.000+00:00');
-      expect(result.joined, '2020-10-15T06:38:00.000+00:00');
-      expect(result.confirm, true);
-      expect(result.mfa, true);
-      expect(result.roles, []);
-    });
+            expect(result.$id, '5e5ea5c16897e');
+                  expect(result.$createdAt, '2020-10-15T06:38:00.000+00:00');
+                  expect(result.$updatedAt, '2020-10-15T06:38:00.000+00:00');
+                  expect(result.userId, '5e5ea5c16897e');
+                  expect(result.userName, 'John Doe');
+                  expect(result.userEmail, 'john@appwrite.io');
+                  expect(result.userPhone, '+1 555 555 5555');
+                  expect(result.teamId, '5e5ea5c16897e');
+                  expect(result.teamName, 'VIP');
+                  expect(result.invited, '2020-10-15T06:38:00.000+00:00');
+                  expect(result.joined, '2020-10-15T06:38:00.000+00:00');
+                  expect(result.confirm, true);
+                  expect(result.mfa, true);
+                  expect(result.roles, []);
+          });
   });
 }