+++ test/src/realtime_subscription_test.dart
@@ -7,9 +7,14 @@ void main() {
   group('RealtimeSubscription', () {
     final mockStream = StreamController<RealtimeMessage>.broadcast();
     final mockCloseFunction = () async {};
+    final mockUnsubscribeFunction = () async {};
+    Future<void> mockUpdateFunction(
+        {List<Object>? channels, List<String>? queries}) async {}
     final subscription = RealtimeSubscription(
         controller: mockStream,
         close: mockCloseFunction,
+        unsubscribe: mockUnsubscribeFunction,
+        update: mockUpdateFunction,
         channels: ['documents'],
         queries: const []);
 
@@ -17,6 +22,8 @@ void main() {
       expect(subscription.controller, equals(mockStream));
       expect(subscription.stream, equals(mockStream.stream));
       expect(subscription.close, equals(mockCloseFunction));
+      expect(subscription.unsubscribe, equals(mockUnsubscribeFunction));
+      expect(subscription.update, equals(mockUpdateFunction));
     });
   });
 }